+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Nothing at the moment
 - feat!: Handle input device connection/disconnection at runtime (#38)
 - feat: Update lv_drivers to git master (2023-03-11)
 - feat: Update lvgl to v8.3 (2023-03-08)
+- feat!: Deprecate -c CLI option and add support for reading from /etc/unl0kr.conf.d/
 
 ## 0.3.0 (2022-11-13)