+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Nothing at the moment
 - feat: Update lv_drivers to git master (2023-03-11)
 - feat: Update lvgl to v8.3 (2023-03-08)
 - feat!: Deprecate -c CLI option and add support for reading from /etc/unl0kr.conf.d/
+- fix: Print correct version in logs
 
 ## 0.3.0 (2022-11-13)